requirements.txt
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
python-telegram-bot==20.7
aiohttp==3.9.1
websockets==12.0
openai==1.6.1
pyttsx3==2.90
playsound==1.3.0
pyautogui==0.9.54
